+++ src/app/user/create-account/page.tsx
@@ -1,9 +1,10 @@
-// Purpose: Allow an Organizer to create an account
+// Purpose: Allow a user to create an account
 
 // Functionality:
-// - Organizer Inputs (update as needed): Name, Email, Password Image (optional)
+// - User Inputs (update as needed): First name, last name, Email, username, Password, Image (optional)
 // - Welcome Modal on submission
 // - Redirect to Profile on completion
+
 'use client'
 import React from 'react'
 import Link from 'next/link'
@@ -11,7 +12,7 @@ import { useRouter } from 'next/navigation'
 import Image from 'next/image'
 import ticket from '../../../utils/images/ticket.jpg'
 
-export default function CreateOrganizerAccountForm() {
+export default function CreateCustomerAccountForm() {
   const router = useRouter()
   // Handles the submit event on form submit.
   const handleSubmit = async (event: any) => {
@@ -20,15 +21,17 @@ export default function CreateOrganizerAccountForm() {
 
     // Get data from the form.
     const data = {
-      organizerName: event.target.organizerName.value,
+      firstName: event.target.firstName.value,
+      lastName: event.target.lastName.value,
       email: event.target.email.value,
+      username: event.target.username.value,
       password: event.target.password.value,
     }
 
     // // Send the data to the server in JSON format.
     // const JSONdata = JSON.stringify(data)
 
-    alert(`Account created for ${data.organizerName}`)
+    alert(`Hi ${data.firstName}`)
 
     router.push('/user/profile')
 
@@ -57,7 +60,7 @@ export default function CreateOrganizerAccountForm() {
   }
   return (
     <div className="min-h-screen flex items-center justify-center py-12 px-4 sm:px-6 lg:px-8 ">
-      <div className="bg-white  max-w-md w-full space-y-8 text-dark border-solid border-2 border-primary ">
+      <div className="bg-white max-w-md w-full space-y-8 text-dark border-solid border-2 border-primary">
         <div>
           <Image
             className="mx-auto h-12 w-auto mt-4"
@@ -65,31 +68,34 @@ export default function CreateOrganizerAccountForm() {
             alt="ticket logo"
           ></Image>
 
-          <h2 className="mt-6 text-center text-3xl font-extrabold">
-            Create your EventPro Organizer Account
+          <h2 className="mt-6 mx-4 text-center text-3xl font-extrabold">
+            Create your EventPro Account
           </h2>
         </div>
         <div className="max-w-md w-full space-y-8">
           <form
+            className="shadow-md rounded px-8 pt-6 pb-8"
             onSubmit={handleSubmit}
-            className=" shadow-md rounded px-8 pt-6 pb-8  "
           >
             <div className="mb-4">
               <label
-                htmlFor="organizerName"
                 className="block text-dark text-sm font-bold mb-2"
+                htmlFor="name"
               >
-                Organizer Name
+                Name
               </label>
               <input
                 className="shadow appearance-none border rounded border-accent w-full py-2 px-3 text-dark leading-tight focus:outline-secondary focus:shadow-outline placeholder-accent"
-                id="organizerName"
+                placeholder="Name"
                 type="text"
-                placeholder="Organizer Name"
-                name="organizerName"
+                id="name"
+                name="name"
                 required
               />
             </div>
+
+           
+
             <div className="mb-4">
               <label
                 className="block text-dark text-sm font-bold mb-2"
@@ -100,9 +106,26 @@ export default function CreateOrganizerAccountForm() {
               <input
                 className="shadow appearance-none border rounded border-accent w-full py-2 px-3 text-dark leading-tight focus:outline-secondary focus:shadow-outline placeholder-accent"
                 placeholder="Email"
+                name="email"
+                required
                 type="email"
                 id="email"
-                name="email"
+              />
+            </div>
+
+            <div className="mb-4">
+              <label
+                className="block text-dark text-sm font-bold mb-2"
+                htmlFor="username"
+              >
+                Username
+              </label>
+              <input
+                className="shadow appearance-none border rounded border-accent w-full py-2 px-3 text-dark leading-tight focus:outline-secondary focus:shadow-outline placeholder-accent"
+                placeholder="Username"
+                type="text"
+                id="username"
+                name="username"
                 required
               />
             </div>
